requirements.txt
Django>=2.0.0
django-cruds>=1.0
rules==1.3.0
django-tables2>=1.20.0
django-appconf
django-filter
rules

# Additional requirements go here
